agentInfoWrapper
