+++ tests/graphai/http_test_bypass.ts
@@ -1,7 +1,7 @@
 import { AgentFunction } from "@/graphai";
 
 import { graphDataTestRunner } from "~/utils/runner";
-import { getAgentInfo } from "@/utils/test_utils";
+import { agentInfoWrapper } from "@/utils/utils";
 
 import test from "node:test";
 import assert from "node:assert";
@@ -52,7 +52,7 @@ test("test bypass1", async () => {
       },
     },
   };
-  const result = await graphDataTestRunner(__filename, graph_data, { httpAgent: getAgentInfo(httpAgent) });
+  const result = await graphDataTestRunner(__filename, graph_data, { httpAgent: agentInfoWrapper(httpAgent) });
   assert.deepStrictEqual(result, {
     bypassAgent2: [
       {
@@ -113,7 +113,7 @@ test("test bypass2", async () => {
       },
     },
   };
-  const result = await graphDataTestRunner(__filename, graph_data, { httpAgent: getAgentInfo(httpAgent) });
+  const result = await graphDataTestRunner(__filename, graph_data, { httpAgent: agentInfoWrapper(httpAgent) });
   // console.log(result);
   assert.deepStrictEqual(result, {
     echo: { message: ["hello", "hello"] },
@@ -182,7 +182,7 @@ test("test bypass3", async () => {
       },
     },
   };
-  const result = await graphDataTestRunner("http_test_bypass_3", graph_data, { httpAgent: getAgentInfo(httpAgent) });
+  const result = await graphDataTestRunner("http_test_bypass_3", graph_data, { httpAgent: agentInfoWrapper(httpAgent) });
   // console.log(result);
   assert.deepStrictEqual(result, {
     echo: { message: ["hello", "hello"] },
@@ -245,7 +245,7 @@ test("test bypass4", async () => {
       },
     },
   };
-  const result = await graphDataTestRunner(__filename, graph_data, { httpAgent: getAgentInfo(httpAgent) });
+  const result = await graphDataTestRunner(__filename, graph_data, { httpAgent: agentInfoWrapper(httpAgent) });
   assert.deepStrictEqual(result, {
     echo: { message: ["hello", "hello"] },
     mapNode: {
