+++ tests/graphai/test_fork.ts
@@ -2,7 +2,7 @@ import { AgentFunction } from "@/graphai";
 
 import { graphDataTestRunner } from "~/utils/runner";
 import { defaultTestAgents } from "@/utils/test_agents";
-import { getAgentInfo } from "@/utils/test_utils";
+import { agentInfoWrapper } from "@/utils/utils";
 
 import test from "node:test";
 import assert from "node:assert";
@@ -72,7 +72,7 @@ test("test fork 1", async () => {
   };
 
   const result = await graphDataTestRunner(__filename, forkGraph, {
-    testAgent1a: getAgentInfo(testAgent1a),
+    testAgent1a: agentInfoWrapper(testAgent1a),
     ...defaultTestAgents,
   });
   // console.log(JSON.stringify(result, null, "  "));
@@ -148,7 +148,7 @@ test("test fork 2", async () => {
   };
 
   const result = await graphDataTestRunner(__filename, forkGraph, {
-    testAgent1: getAgentInfo(testAgent1),
+    testAgent1: agentInfoWrapper(testAgent1),
   });
   // console.log(JSON.stringify(result, null, "  "));
   assert.deepStrictEqual(result, {
