+++ tests/graphai/test_loop_2.ts
@@ -2,7 +2,7 @@ import { AgentFunction } from "@/graphai";
 import { graphDataTestRunner } from "~/utils/runner";
 
 import { defaultTestAgents } from "@/utils/test_agents";
-import { getAgentInfo } from "@/utils/test_utils";
+import { agentInfoWrapper } from "@/utils/utils";
 
 import test from "node:test";
 import assert from "node:assert";
@@ -31,7 +31,7 @@ const counterAgent: AgentFunction = async ({ inputs }) => {
 
 test("test counter", async () => {
   const result = await graphDataTestRunner(__filename, graphdata_counter, {
-    counterAgent: getAgentInfo(counterAgent),
+    counterAgent: agentInfoWrapper(counterAgent),
   });
   assert.deepStrictEqual(result, { data: { v: 9 }, counter: { v: 10 } });
 });
@@ -61,7 +61,7 @@ test("test counter2", async () => {
 
   const result = await graphDataTestRunner(__filename, nested_graphdata, {
     ...defaultTestAgents,
-    counterAgent: getAgentInfo(counterAgent),
+    counterAgent: agentInfoWrapper(counterAgent),
   });
   assert.deepStrictEqual(result, { workingMemory: { v: 90 }, nested1: { counter: { v: 100 } } });
 });
@@ -100,7 +100,7 @@ test("test counter3", async () => {
 
   const result = await graphDataTestRunner(__filename, nested_graphdata, {
     ...defaultTestAgents,
-    counterAgent: getAgentInfo(counterAgent),
+    counterAgent: agentInfoWrapper(counterAgent),
   });
   assert.deepStrictEqual(result, {
     workingMemory: { v: 10220 },
